bugfix(ET-325): change ContactCardType ToSql implementation to correctly cast into u8 rather than isize

+++ crypto/crypto-account/src/contacts/cards.rs
@@ -33,7 +33,7 @@ pub enum ContactCardType {
 
 impl ToSql for ContactCardType {
     fn to_sql(&self) -> Result<ToSqlOutput, rusqlite::Error> {
-        Ok(ToSqlOutput::from(*self as isize))
+        Ok(ToSqlOutput::from(*self as u8))
     }
 }
